OMA-52 Update educandu to v0.28.0
