+++ src/bundles/home-page-template.less
@@ -4,7 +4,3 @@
   flex-direction: column;
   margin-bottom: 10px;
 }
-
-.HomePageTemplate-info {
-  .m-content;
-}
