+++ src/bundles/page-footer.less
@@ -1,9 +1,9 @@
 .PageFooter {
+  .m-page-content;
   flex: none;
   display: flex;
   align-items: center;
   justify-content: center;
-  padding: @edu-content-padding;
   position: relative;
   z-index: @edu-z-index-base-layer;
 }
