+++ src/bundles/page-template.js
@@ -1,3 +1,4 @@
+import { Alert } from 'antd';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import React, { useState } from 'react';
@@ -26,11 +27,30 @@ function PageTemplate({ children, fullScreen, alerts }) {
     'PageTemplate-content--fullScreen': fullScreen
   });
 
+  const renderAlert = (alert, index) => {
+    const shouldRenderAlert = !fullScreen || alert.showInFullScreen;
+    if (!shouldRenderAlert) {
+      return null;
+    }
+
+    return (
+      <Alert
+        key={index}
+        message={alert.message}
+        type={alert.type || 'info'}
+        banner
+        closable={alert.closable || false}
+        onClose={alert.onClose || (() => { })}
+        />
+    );
+  };
+
   return (
     <div className="PageTemplate">
       <PageHeader fullScreen={fullScreen} alerts={alerts} onUiLanguageClick={handleUiLanguageClick} />
       <main className={contentAreaClasses}>
         <div className={contentClasses}>
+          {!!alerts?.length && <div className="PageTemplate-contentAlerts">{alerts.map(renderAlert)}</div>}
           {children}
         </div>
       </main>
