+++ src/bundles/page-template.less
@@ -10,6 +10,7 @@
 }
 
 .PageTemplate-contentArea {
+  .m-page-content;
   flex: 1 0 0%;
   display: flex;
   align-items: flex-start;
@@ -35,4 +36,8 @@
     padding-top: 20vh;
     padding-bottom: 30vh;
   }
+}
+
+.PageTemplate-contentAlerts {
+  padding-bottom: @edu-content-padding;
 }
