+++ src/main.less
@@ -9,7 +9,7 @@
 @import url('./bundles/home-page-template.less');
 @import url('./bundles/home-page-illustration.less');
 
-@oma-blue: #002F65;
+@oma-blue: #6D8BB1;
 
 // Override educandu styles:
 @edu-font-family: 'Lato', sans-serif;
@@ -19,3 +19,7 @@
 @edu-success-color: #369911;
 @edu-error-color: #b11200;
 @edu-warning-color: #fa8c16;
+
+@edu-link-color: hardlight(@edu-primary-color, @grey-7);
+@edu-link-hover-color: darken(@edu-link-color, 10%);
+@edu-link-active-color: darken(@edu-link-color, 10%);
